==> 3_text-preprocessing.ipynb <==
import pandas as pd
import numpy as np
import preprocessor as p
import requests

from tqdm.notebook import tqdm
tqdm.pandas()

import warnings
warnings.filterwarnings('ignore')

pd.set_option('display.max_colwidth', None)

raw = pd.read_csv('data/data.csv')
raw.shape

raw.head()

raw.info()

# 1. Data Preprocessing
---

## 1.1 Filter Tweets

Drop retweeted Tweets to diminish duplication

raw['reference_type'].value_counts()

We will drop 66528 retweeted Tweets

df = raw[raw['reference_type'] != "['retweeted']"].reset_index(drop=True)

df.shape

Drop duplicated tweets and keep the first occurance

duplicate = df[df.duplicated('text')]
print('Total duplicated tweets:',duplicate.shape[0])

df.drop_duplicates(subset=['text'], inplace=True)

df.shape

## 1.2 Text Preprocessing

# Normalisasi Kata
url1 = (r'https://raw.githubusercontent.com/ksnugroho/klasifikasi-spam-sms/master/data/key_norm.csv')
url2 = (r'https://raw.githubusercontent.com/nasalsabila/kamus-alay/master/colloquial-indonesian-lexicon.csv')
url3 = (r'https://raw.githubusercontent.com/okkyibrohim/id-abusive-language-detection/master/kamusalay.csv')
url4 = (r'https://raw.githubusercontent.com/louisowen6/NLP_bahasa_resources/master/combined_slang_words.txt')
r = requests.get(url4)

kamus_normalisasi = pd.read_csv(url1)
kamus_alay1 = pd.read_csv(url2)
kamus_alay2 = pd.read_csv(url3, names=['slang','formal'])
kamus_slang = pd.json_normalize(r.json()).transpose().reset_index()
kamus_slang.columns = ['slang','formal']

def normalisasi(text):
  text = ' '.join([kamus_normalisasi[kamus_normalisasi['singkat'] == word]['hasil'].values[0] if (kamus_normalisasi['singkat'] == word).any() else word for word in text.split()])
  text = ' '.join([kamus_alay1[kamus_alay1['slang'] == word]['formal'].values[0] if (kamus_alay1['slang'] == word).any() else word for word in text.split()])
  text = ' '.join([kamus_alay2[kamus_alay2['slang'] == word]['formal'].values[0] if (kamus_alay2['slang'] == word).any() else word for word in text.split()])
  text = ' '.join([kamus_slang[kamus_slang['slang'] == word]['formal'].values[0] if (kamus_slang['slang'] == word).any() else word for word in text.split()])
  return text

def normalisasi_slang(text):
  text = ' '.join([kamus_alay1[kamus_alay1['slang'] == word]['formal'].values[0] if (kamus_alay1['slang'] == word).any() else word for word in text.split()])
  text = ' '.join([kamus_alay2[kamus_alay2['slang'] == word]['formal'].values[0] if (kamus_alay2['slang'] == word).any() else word for word in text.split()])
  text = ' '.join([kamus_slang[kamus_slang['slang'] == word]['formal'].values[0] if (kamus_slang['slang'] == word).any() else word for word in text.split()])
  return text

# Pipeline
def preprocessing(text):
  text = text.replace('\n', ' ').replace('\r', '') # replace new line into space
  text = text.replace('&amp;', ' ')
  p.set_options(p.OPT.URL, p.OPT.MENTION, p.OPT.HASHTAG, p.OPT.RESERVED)
  text = p.clean(text)
  text = normalisasi(text)
  #text = normalisasi_slang(text)
  return text

# Start preprocessing
df['text_cleaned'] = df['text'].progress_apply(preprocessing)

df[['text','text_cleaned']].sample(10)

Remove the duplicated cleaned tweets again

duplicate2 = df[df.duplicated('text_cleaned')]
print('Total duplicated tweets:',duplicate2.shape[0])

df.drop_duplicates(subset=['text_cleaned'], inplace=True)

df.shape

# Save the prepocessing result
df.to_csv('data/data_preprocessed.csv', index=False)
